Made changes to restrict access to editing and delete in backend only to the user that added the recommendation

+++ src/main/java/fi/haagahelia/coolreads/controller/ReadingRecommendationController.java
@@ -3,6 +3,7 @@ package fi.haagahelia.coolreads.controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.server.ResponseStatusException;
 
 import fi.haagahelia.coolreads.dto.RecommendationDto;
 import fi.haagahelia.coolreads.model.AppUser;
@@ -80,28 +82,47 @@ public class ReadingRecommendationController {
 		return "recommendationlist";
 	}
 
+	
 	@GetMapping("/edit/{id}")
-	public String editReadingRecommendation(@PathVariable("id") Long id, Model model) {
+	public String editReadingRecommendation(@PathVariable("id") Long id, Model model, @AuthenticationPrincipal UserDetails userDetails) {
+		
 		model.addAttribute("categories", categoryRepository.findAll());
 		Recommendation recommendation = readingRepository.findById(id).orElse(null);
+		if (userDetails.getUsername().equalsIgnoreCase(recommendation.getUser().getUsername())) {
+			if (recommendation != null) {
+				model.addAttribute("recommendation", recommendation);
+				return "editrecommendation";
+			}
+		}
+		
+		return "redirect:/";
 
-		if (recommendation != null) {
+		/*if (recommendation != null) {
 			model.addAttribute("recommendation", recommendation);
 			return "editrecommendation";
 		} else {
 			return "redirect:/";
-		}
+		} */
 	}
 
 	@PostMapping("/saveEditedReadingRecommendation")
-	public String saveEditedBook(@ModelAttribute Recommendation recommendationForm) {
-		readingRepository.save(recommendationForm);
-		return "redirect:/";
+	public String saveEditedBook(@ModelAttribute Recommendation recommendationForm, @AuthenticationPrincipal UserDetails userDetails) {
+		if (userDetails.getUsername().equalsIgnoreCase(recommendationForm.getUser().getUsername())) {
+			readingRepository.save(recommendationForm);
+			return "redirect:/";
+		} else {
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized user");
+		}
+		
 	}
 
 	@PostMapping("/delete/{id}")
-	public String deleteRecommendation(@PathVariable("id") Long id, Model model) {
-		readingRepository.deleteById(id);
+	public String deleteRecommendation(@PathVariable("id") Long id, Model model, @AuthenticationPrincipal UserDetails userDetails) {
+		Recommendation recommendation = readingRepository.findById(id).orElse(null);
+		if (userDetails.getUsername().equalsIgnoreCase(recommendation.getUser().getUsername())) {
+			readingRepository.deleteById(id);
+		}
+		
 		return "redirect:/";
 	}
 }
